[REFACTOR] Balance 패키지 이동
