+++ src/main/java/kr/hhplus/be/ecommerce/interfaces/balance/api/BalanceController.java
@@ -1,8 +1,8 @@
-package kr.hhplus.be.ecommerce.interfaces.balance;
+package kr.hhplus.be.ecommerce.interfaces.balance.api;
 
 import jakarta.validation.Valid;
-import kr.hhplus.be.ecommerce.application.balance.BalanceFacade;
-import kr.hhplus.be.ecommerce.application.balance.BalanceResult;
+import kr.hhplus.be.ecommerce.domain.balance.BalanceInfo;
+import kr.hhplus.be.ecommerce.domain.balance.BalanceService;
 import kr.hhplus.be.ecommerce.interfaces.ApiResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.*;
@@ -11,18 +11,18 @@ import org.springframework.web.bind.annotation.*;
 @RequiredArgsConstructor
 public class BalanceController {
 
-    private final BalanceFacade balanceFacade;
+    private final BalanceService balanceService;
 
     @GetMapping("/api/v1/users/{id}/balance")
     public ApiResponse<BalanceResponse.Balance> getBalance(@PathVariable("id") Long id) {
-        BalanceResult.Balance balance = balanceFacade.getBalance(id);
+        BalanceInfo.Balance balance = balanceService.getBalance(id);
         return ApiResponse.success(BalanceResponse.Balance.of(balance));
     }
 
     @PostMapping("/api/v1/users/{id}/balance/charge")
     public ApiResponse<Void> chargeBalance(@PathVariable("id") Long id,
                                            @Valid @RequestBody BalanceRequest.Charge request) {
-        balanceFacade.chargeBalance(request.toCriteria(id));
+        balanceService.chargeBalance(request.toCommand(id));
         return ApiResponse.success();
     }
 }
